back/requirements.txt
python-dotenv
fastapi==0.75.1
uvicorn[standard]
motor==2.5.1
pymongo==3.12.3
